refactoring folder structure based on strategy pattern
