+++ javadoc/model/strategy/DiceStrategy.html
@@ -77,7 +77,7 @@ loadScripts(document, 'script');</script>
 <h1 title="Class DiceStrategy" class="title">Class DiceStrategy</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree">java.lang.Object
-<div class="inheritance">model.strategy.DiceStrategy</div>
+<div class="inheritance">model.strategy.game.DiceStrategy</div>
 </div>
 <section class="description">
 <dl class="notes">
@@ -149,7 +149,7 @@ int&nbsp;p_Armies)</code></th>
 <h3 id="methods.inherited.from.class.java.lang.Object">Methods inherited from class&nbsp;java.lang.Object</h3>
 <code>clone, equals, finalize, getClass, hashCode, notify, notifyAll, toString, wait, wait, wait</code></div>
 <div class="inherited-list">
-<h3 id="methods.inherited.from.class.model.strategy.GameStrategy">Methods inherited from interface&nbsp;model.strategy.<a href="GameStrategy.html" title="interface in model.strategy">GameStrategy</a></h3>
+<h3 id="methods.inherited.from.class.model.strategy.game.GameStrategy">Methods inherited from interface&nbsp;model.strategy.<a href="GameStrategy.html" title="interface in model.strategy">GameStrategy</a></h3>
 <code><a href="GameStrategy.html#makeMeKing(model.Player,model.Country)">makeMeKing</a></code></div>
 </section>
 </li>
