+++ src/main/java/model/strategy/game/DefaultStrategy.java
@@ -1,4 +1,4 @@
-package model.strategy;
+package model.strategy.game;
 
 import model.Card;
 import model.Country;
@@ -6,14 +6,13 @@ import model.GameSettings;
 import model.Player;
 import utils.logger.LogEntryBuffer;
 
-import java.util.stream.IntStream;
-
 /**
- * Class holding the Dice strategy for advance logic
+ * Class holding the default strategy of the advance order.
  *
  * @author Madhuvanthi Hemanathan
  */
-public class DiceStrategy implements GameStrategy {
+public class DefaultStrategy implements GameStrategy {
+
     /**
      * Game settings object
      */
@@ -37,33 +36,30 @@ public class DiceStrategy implements GameStrategy {
     public boolean attack(Player p_Player, Country p_From, Country p_To, int p_Armies) {
         try {
             p_From.depleteArmies(p_Armies);
-            int l_AttackerKills = (int) IntStream.range(0, p_Armies).boxed().filter((p_integer) -> Math.random() <= SETTINGS.ATTACKER_PROBABILITY).count();
-            int l_DefenderKills = (int) IntStream.range(0, p_To.getArmies()).boxed().filter(p_integer -> Math.random() <= SETTINGS.DEFENDER_PROBABILITY).count();
+            int l_attackerKills = (int) Math.round(p_Armies * SETTINGS.ATTACKER_PROBABILITY);
+            int l_defenderKills = (int) Math.round(p_To.getArmies() * SETTINGS.DEFENDER_PROBABILITY);
 
-            int l_ArmiesLeftAttacker = p_Armies - l_DefenderKills;
-            int l_ArmiesLeftDefender = p_To.getArmies() - l_AttackerKills;
-            if (l_ArmiesLeftAttacker > 0 && l_ArmiesLeftDefender <= 0) {
-                p_To.setArmies(l_ArmiesLeftAttacker);
+            int l_armiesLeftAttacker = p_Armies - l_defenderKills;
+            int l_armiesLeftDefender = p_To.getArmies() - l_attackerKills;
+            if (l_armiesLeftAttacker > 0 && l_armiesLeftDefender <= 0) {
+                p_To.setArmies(l_armiesLeftAttacker);
                 makeMeKing(p_Player, p_To);
                 d_Logger.log(String.format("Attacker : %s (%s) won",p_Player.getName(),p_From.getName()));
-                // Assign a card
+                //Assign power card to king
                 Card l_AssignedCard = new Card();
                 p_Player.addPlayerCard(l_AssignedCard);
                 d_Logger.log("Attacker: " + p_Player.getName() + " received a card: " + l_AssignedCard.getCardType().toString());
-                d_Logger.log(String.format("Since won, left out %s (Attacker) armies %s moved to %s.",p_From.getName(),l_ArmiesLeftAttacker,p_To.getName()));
-
+                d_Logger.log(String.format("Since won, left out %s (Attacker) armies %s moved to %s.",p_From.getName(),l_armiesLeftAttacker,p_To.getName()));
             } else {
-                p_From.deployArmies(l_ArmiesLeftAttacker);
-                p_To.setArmies(l_ArmiesLeftDefender);
-                d_Logger.log("Attacker : " + p_Player.getName() + " lost.");
-                d_Logger.log("Remaining attacker's armies: " + p_From.getArmies());
-                d_Logger.log("Remaining defender's armies: " + p_To.getArmies());
+                p_From.deployArmies(l_armiesLeftAttacker);
+                p_To.setArmies(l_armiesLeftDefender);
+                d_Logger.log(String.format("Attacker : %s (%s) lost" ,p_Player.getName(),p_From.getName()));
+                d_Logger.log(String.format("Remaining armies of %s (Attacker) in attack: %s ", p_From.getName(), l_armiesLeftAttacker));
+                d_Logger.log(String.format("Remaining armies of %s (Defender) in attack: %s ", p_To.getName(), l_armiesLeftDefender));
             }
             return true;
         } catch (Exception p_Exception) {
             return false;
         }
-
     }
-
 }
